Add logic for checking indices in delete function

+++ SOAP.py
@@ -114,6 +114,9 @@ def delete(table, cursor):
             user_choice += 1
             row_to_delete.append(item[0])
         user_choice = int(input("Pick a row to delete: "))
+        if user_choice < 1 or user_choice > len(row_to_delete):
+            print("Error: Index out of range.")
+            return
         try:
             cursor.execute("delete from Agency where Agency_ID=(?)", [row_to_delete[user_choice - 1]])
         except sqlite3.Error as e:
@@ -127,6 +130,9 @@ def delete(table, cursor):
             user_choice += 1
             row_to_delete.append(item[0])
         user_choice = int(input("Pick a row to delete: "))
+        if user_choice < 1 or user_choice > len(row_to_delete):
+            print("Error: Index out of range.")
+            return
         try:
             cursor.execute("delete from Office where Office_Name =(?)", [row_to_delete[user_choice - 1]])
         except sqlite3.Error as e:
@@ -140,6 +146,9 @@ def delete(table, cursor):
             user_choice += 1
             row_to_delete.append(item[0])
         user_choice = int(input("Pick a row to delete: "))
+        if user_choice < 1 or user_choice > len(row_to_delete):
+            print("Error: Index out of range.")
+            return
         try:
             cursor.execute("delete from Rental_Agreement where Agreement_ID =(?)", [row_to_delete[user_choice - 1]])
         except sqlite3.Error as e:
@@ -153,6 +162,9 @@ def delete(table, cursor):
             user_choice += 1
             row_to_delete.append([item[0], item[1]])
         user_choice = int(input("Pick a row to delete: "))
+        if user_choice < 1 or user_choice > len(row_to_delete):
+            print("Error: Index out of range.")
+            return
         try:
             cursor.execute("delete from Agrees where Agreement_ID =(?) and Agency_ID = (?)", 
                 [row_to_delete[user_choice - 1][0], row_to_delete[user_choice - 1][1]])
